Refactor circuit sharing logic into dedicated module
