+++ src/main.js
@@ -1,6 +1,6 @@
 // Entry point module coordinating Bitwiser features.
 // Placeholder imports ensure upcoming modules can hook into the bootstrap flow.
-import { initializeAuth, ensureDriveAuth } from './modules/auth.js';
+import { initializeAuth } from './modules/auth.js';
 import {
   getUsername,
   setUsername,
@@ -35,6 +35,20 @@ import {
 import * as levelsModule from './modules/levels.js';
 import * as uiModule from './modules/ui.js';
 import { openHintModal, initializeHintUI } from './modules/hints.js';
+import {
+  initializeCircuitShare,
+  updateSaveProgress,
+  handleGifModalClose,
+  handleGifSaveClick,
+  handleGifCopyClick,
+  handleGifShareClick,
+  handleGIFExport,
+  handleSaveCircuitClick,
+  saveCircuit,
+  openSavedModal,
+  closeSavedModal,
+  loadGifFromDB
+} from './modules/circuitShare.js';
 import {
   setupNavigation,
   lockOrientationLandscape,
@@ -130,63 +144,36 @@ const gifLoadingText = document.getElementById('gifLoadingText');
 const saveProgressContainer = document.getElementById('saveProgressContainer');
 const saveProgressBar = document.getElementById('saveProgressBar');
 
-function updateSaveProgress(percent) {
-  if (saveProgressBar) {
-    saveProgressBar.style.width = `${percent}%`;
-  }
-}
-let currentGifBlob = null;
-let currentGifUrl = null;
+initializeCircuitShare({
+  elements: {
+    gifModal,
+    gifPreview,
+    gifLoadingModal,
+    gifLoadingText,
+    saveProgressContainer,
+    saveProgressBar,
+    savedModal: document.getElementById('savedModal'),
+    savedList: document.getElementById('savedList')
+  },
+  translate: typeof t === 'function' ? t : key => key,
+  alert,
+  confirm,
+  getCurrentCustomProblem: () => currentCustomProblem,
+  getCurrentCustomProblemKey: () => currentCustomProblemKey,
+  onLastSavedKeyChange: key => { lastSavedKey = key; }
+});
+
 if (closeGifModalBtn) {
-  closeGifModalBtn.addEventListener('click', () => {
-    if (gifModal) gifModal.style.display = 'none';
-    if (gifPreview && currentGifUrl) {
-      URL.revokeObjectURL(currentGifUrl);
-      gifPreview.src = '';
-      currentGifUrl = null;
-    }
-    currentGifBlob = null;
-  });
+  closeGifModalBtn.addEventListener('click', handleGifModalClose);
 }
 if (saveGifBtn) {
-  saveGifBtn.addEventListener('click', () => {
-    if (!currentGifUrl) return;
-    const link = document.createElement('a');
-    link.href = currentGifUrl;
-    link.download = 'circuit.gif';
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  });
+  saveGifBtn.addEventListener('click', handleGifSaveClick);
 }
 if (copyGifBtn) {
-  copyGifBtn.addEventListener('click', async () => {
-    if (!currentGifBlob) return;
-    try {
-      await navigator.clipboard.write([
-        new ClipboardItem({ [currentGifBlob.type]: currentGifBlob })
-      ]);
-      alert('이미지가 클립보드에 복사되었습니다.');
-    } catch (err) {
-      console.error(err);
-      alert('이미지 복사에 실패했습니다.');
-    }
-  });
+  copyGifBtn.addEventListener('click', handleGifCopyClick);
 }
 if (shareGifBtn) {
-  shareGifBtn.addEventListener('click', async () => {
-    if (!currentGifBlob) return;
-    const file = new File([currentGifBlob], 'circuit.gif', { type: 'image/gif' });
-    if (navigator.share && navigator.canShare && navigator.canShare({ files: [file] })) {
-      try {
-        await navigator.share({ files: [file], title: 'Bitwiser GIF' });
-      } catch (err) {
-        console.error(err);
-      }
-    } else {
-      alert('이 브라우저에서는 공유하기를 지원하지 않습니다.');
-    }
-  });
+  shareGifBtn.addEventListener('click', handleGifShareClick);
 }
 
 // 초기 로딩 관련
@@ -1362,279 +1349,17 @@ function mergeProgress(oldName, newName) {
 // 1) 모달과 버튼 요소 참조
 const viewSavedBtn = document.getElementById('viewSavedBtn');
 const saveCircuitBtn = document.getElementById('saveCircuitBtn');
-const savedModal = document.getElementById('savedModal');
-const closeSavedModal = document.getElementById('closeSavedModal');
+const closeSavedModalBtn = document.getElementById('closeSavedModal');
 
-// Version tag for canvas-based circuit saves
-const CURRENT_CIRCUIT_VERSION = 2;
-
-// Google Drive file helpers
-async function uploadFileToAppData(name, blob, mimeType) {
-  await ensureDriveAuth();
-  const token = gapi.client.getToken().access_token;
-  const metadata = {
-    name,
-    parents: ['appDataFolder'],
-    mimeType
-  };
-  const form = new FormData();
-  form.append('metadata', new Blob([JSON.stringify(metadata)], { type: 'application/json' }));
-  form.append('file', blob);
-  await fetch('https://www.googleapis.com/upload/drive/v3/files?uploadType=multipart&fields=id', {
-    method: 'POST',
-    headers: new Headers({ 'Authorization': 'Bearer ' + token }),
-    body: form
-  });
+if (saveCircuitBtn) {
+  saveCircuitBtn.addEventListener('click', handleSaveCircuitClick);
 }
 
-async function downloadFileFromAppData(name) {
-  await ensureDriveAuth();
-  const list = await gapi.client.drive.files.list({
-    spaces: 'appDataFolder',
-    fields: 'files(id, name)',
-    q: `name='${name}'`
-  });
-  const file = list.result.files[0];
-  if (!file) return null;
-  const token = gapi.client.getToken().access_token;
-  const res = await fetch(`https://www.googleapis.com/drive/v3/files/${file.id}?alt=media`, {
-    headers: new Headers({ 'Authorization': 'Bearer ' + token })
-  });
-  return await res.blob();
+if (viewSavedBtn) {
+  viewSavedBtn.addEventListener('click', openSavedModal);
 }
-
-async function deleteFileFromAppData(name) {
-  await ensureDriveAuth();
-  const list = await gapi.client.drive.files.list({
-    spaces: 'appDataFolder',
-    fields: 'files(id, name)',
-    q: `name='${name}'`
-  });
-  const file = list.result.files[0];
-  if (file) await gapi.client.drive.files.delete({ fileId: file.id });
-}
-
-async function saveGifToDB(key, blob) {
-  return uploadFileToAppData(`${key}.gif`, blob, 'image/gif');
-}
-
-async function loadGifFromDB(key) {
-  return downloadFileFromAppData(`${key}.gif`);
-}
-
-async function deleteGifFromDB(key) {
-  return deleteFileFromAppData(`${key}.gif`);
-}
-
-saveCircuitBtn.addEventListener('click', async () => {
-  try {
-    await ensureDriveAuth();
-  } catch (e) {
-    alert(e.message);
-    return;
-  }
-  let saveSuccess = false;
-  try {
-    if (gifLoadingModal) {
-      if (gifLoadingText) gifLoadingText.textContent = t('savingCircuit');
-      gifLoadingModal.style.display = 'flex';
-    }
-    if (saveProgressContainer) {
-      saveProgressContainer.style.display = 'block';
-      updateSaveProgress(0);
-    }
-    await saveCircuit(updateSaveProgress);
-    saveSuccess = true;
-  } catch (e) {
-    alert(t('saveFailed').replace('{error}', e));
-  } finally {
-    if (gifLoadingModal) {
-      gifLoadingModal.style.display = 'none';
-      if (gifLoadingText) gifLoadingText.textContent = t('gifLoadingText');
-    }
-    if (saveProgressContainer) {
-      saveProgressContainer.style.display = 'none';
-      updateSaveProgress(0);
-    }
-  }
-  if (saveSuccess) alert(t('circuitSaved'));
-});
-
-// 2) 저장된 회로 키 prefix 계산
-function getSavePrefix() {
-  const level = getCurrentLevel();
-  if (level != null) {
-    return `bit_saved_stage_${String(level).padStart(2, '0')}_`;
-  } else if (currentCustomProblemKey) {
-    return `bit_saved_prob_${currentCustomProblemKey}_`;
-  }
-  return 'bit_saved_';
-}
-
-async function listCircuitJsonFiles() {
-  await ensureDriveAuth();
-  const res = await gapi.client.drive.files.list({
-    spaces: 'appDataFolder',
-    fields: 'files(id, name, createdTime)',
-    q: "name contains '.json'"
-  });
-  return res.result.files || [];
-}
-
-// 3) 리스트 그리기
-async function renderSavedList() {
-  const savedList = document.getElementById('savedList');
-  savedList.innerHTML = `<p>${t('loadingText')}</p>`;
-  try {
-    await ensureDriveAuth();
-  } catch (e) {
-    savedList.innerHTML = `<p>${t('loginRequired')}</p>`;
-    return;
-  }
-  const prefix = getSavePrefix();
-  const files = (await listCircuitJsonFiles())
-    .filter(f => f.name.startsWith(prefix))
-    .sort((a, b) => new Date(b.createdTime) - new Date(a.createdTime));
-  if (!files.length) {
-    savedList.innerHTML = `<p>${t('noCircuits')}</p>`;
-    return;
-  }
-  const items = await Promise.all(files.map(async file => {
-    const key = file.name.replace('.json', '');
-    const [blob, gifBlob] = await Promise.all([
-      downloadFileFromAppData(file.name),
-      loadGifFromDB(key)
-    ]);
-    const text = await blob.text();
-    const data = JSON.parse(text);
-    const item = document.createElement('div');
-    item.className = 'saved-item';
-    const label = data.stageId != null
-      ? `Stage ${String(data.stageId).padStart(2, '0')}`
-      : `Problem ${data.problemTitle || data.problemKey}`;
-
-    const img = document.createElement('img');
-    if (gifBlob) img.src = URL.createObjectURL(gifBlob);
-    img.alt = label;
-    item.appendChild(img);
-
-    const cap = document.createElement('div');
-    cap.className = 'saved-caption';
-    cap.textContent = `${label} — ${new Date(data.timestamp).toLocaleString()}`;
-    item.appendChild(cap);
-
-    item.addEventListener('click', () => {
-      applyCircuitData(data, key);
-      document.getElementById('savedModal').style.display = 'none';
-    });
-
-    const delBtn = document.createElement('button');
-    delBtn.textContent = t('deleteBtn');
-    delBtn.className = 'deleteBtn';
-    delBtn.addEventListener('click', async e => {
-      e.stopPropagation();
-      if (!confirm(t('confirmDelete'))) return;
-      await deleteFileFromAppData(`${key}.json`);
-      await deleteGifFromDB(key);
-      renderSavedList();
-    });
-    item.appendChild(delBtn);
-
-    return item;
-  }));
-  savedList.innerHTML = '';
-  items.forEach(item => savedList.appendChild(item));
-}
-
-// 4) 모달 열기/닫기
-document.getElementById('viewSavedBtn')
-  .addEventListener('click', () => {
-    document.getElementById('savedModal').style.display = 'flex';
-    renderSavedList();
-  });
-document.getElementById('closeSavedModal')
-  .addEventListener('click', () => {
-    document.getElementById('savedModal').style.display = 'none';
-  });
-
-// 5) 회로 불러오는 함수
-async function loadCircuit(key) {
-  try {
-    await ensureDriveAuth();
-  } catch (e) {
-    alert(t('loginRequired'));
-    return;
-  }
-  const blob = await downloadFileFromAppData(`${key}.json`);
-  if (!blob) return alert(t('loadFailedNoData'));
-  const text = await blob.text();
-  const data = JSON.parse(text);
-  applyCircuitData(data, key);
-}
-
-function applyCircuitData(data, key) {
-  if (data.version !== CURRENT_CIRCUIT_VERSION || !data.circuit) {
-    alert(t('incompatibleCircuit'));
-    return;
-  }
-  const circuit = getActiveCircuit();
-  if (!circuit) return;
-  circuit.rows = data.circuit.rows;
-  circuit.cols = data.circuit.cols;
-  circuit.blocks = data.circuit.blocks || {};
-  circuit.wires = data.circuit.wires || {};
-  markCircuitModified();
-  const controller = getActiveController();
-  controller?.syncPaletteWithCircuit?.();
-  controller?.clearSelection?.();
-  if (key) lastSavedKey = key;
-}
-
-
-async function saveCircuit(progressCallback) {
-  const circuit = getActiveCircuit();
-  if (!circuit) throw new Error('No circuit to save');
-
-  await ensureDriveAuth();
-
-  const wireCells = new Set();
-  Object.values(circuit.wires).forEach(w => {
-    w.path.slice(1, -1).forEach(p => wireCells.add(`${p.r},${p.c}`));
-  });
-
-  const data = {
-    version: CURRENT_CIRCUIT_VERSION,
-    stageId: getCurrentLevel(),
-    problemKey: currentCustomProblemKey,
-    problemTitle: currentCustomProblem ? currentCustomProblem.title : undefined,
-    timestamp: new Date().toISOString(),
-    circuit,
-    usedBlocks: Object.keys(circuit.blocks).length,
-    usedWires: wireCells.size
-  };
-
-  const timestampMs = Date.now();
-  const key = `${getSavePrefix()}${timestampMs}`;
-  try {
-    progressCallback && progressCallback(0);
-    const jsonBlob = new Blob([JSON.stringify(data)], { type: 'application/json' });
-    await uploadFileToAppData(`${key}.json`, jsonBlob, 'application/json');
-    progressCallback && progressCallback(33);
-
-    // capture GIF and store in Drive
-    const blob = await new Promise(resolve => captureGIF(resolve));
-    progressCallback && progressCallback(66);
-    await saveGifToDB(key, blob);
-    progressCallback && progressCallback(100);
-
-    console.log(`Circuit saved: ${key}`, data);
-    lastSavedKey = key;
-    return key;
-  } catch (e) {
-    console.error('Circuit save failed:', e);
-    alert('회로 저장 중 오류가 발생했습니다.');
-    throw e;
-  }
+if (closeSavedModalBtn) {
+  closeSavedModalBtn.addEventListener('click', closeSavedModal);
 }
 
 configureLevelModule({ showOverallRanking });
@@ -2172,79 +1897,6 @@ async function gradeProblemCanvas(key, problem) {
   returnBtn.textContent = t('returnToEditBtn');
   gradingArea.appendChild(returnBtn);
   document.getElementById('returnToEditBtn').addEventListener('click', returnToEditScreen);
-}
-
-// ----- GIF 캡처 기능 -----
-
-async function captureGIF(onFinish) {
-  const bgCanvas = document.getElementById('bgCanvas');
-  const contentCanvas = document.getElementById('contentCanvas');
-  if (!bgCanvas || !contentCanvas) return;
-
-  const dpr = window.devicePixelRatio || 1;
-  const totalWidth = bgCanvas.width / dpr;
-  const totalHeight = bgCanvas.height / dpr;
-
-  let gridWidth = totalWidth;
-  let gridHeight = totalHeight;
-  let panelWidth = 0;
-
-  try {
-    const { CELL, GAP } = await import('./canvas/model.js');
-    const controller = getActiveController();
-    const circuit = controller?.circuit || getActiveCircuit();
-    if (circuit) {
-      gridWidth = circuit.cols * (CELL + GAP) + GAP;
-      gridHeight = circuit.rows * (CELL + GAP) + GAP;
-      panelWidth = totalWidth - gridWidth;
-    }
-  } catch (_) {}
-
-  const tempCanvas = document.createElement('canvas');
-  tempCanvas.width = gridWidth;
-  tempCanvas.height = gridHeight;
-  const tempCtx = tempCanvas.getContext('2d');
-
-  const gif = new GIF({ workers: 2, quality: 10, width: gridWidth, height: gridHeight });
-  const totalFrames = 10;
-
-  for (let f = 0; f < totalFrames; f++) {
-    tempCtx.clearRect(0, 0, gridWidth, gridHeight);
-    [bgCanvas, contentCanvas].forEach(c => {
-      tempCtx.drawImage(
-        c,
-        panelWidth * dpr,
-        0,
-        gridWidth * dpr,
-        gridHeight * dpr,
-        0,
-        0,
-        gridWidth,
-        gridHeight
-      );
-    });
-    gif.addFrame(tempCanvas, { delay: 50, copy: true });
-    await new Promise(r => setTimeout(r, 50));
-  }
-
-  gif.on('finished', blob => {
-    if (typeof onFinish === 'function') onFinish(blob);
-  });
-
-  gif.render();
-}
-
-function handleGIFExport() {
-  if (gifLoadingText) gifLoadingText.textContent = t('gifLoadingText');
-  if (gifLoadingModal) gifLoadingModal.style.display = 'flex';
-  captureGIF(blob => {
-    if (gifLoadingModal) gifLoadingModal.style.display = 'none';
-    currentGifBlob = blob;
-    if (currentGifUrl) URL.revokeObjectURL(currentGifUrl);
-    currentGifUrl = URL.createObjectURL(blob);
-    if (gifPreview) gifPreview.src = currentGifUrl;
-    if (gifModal) gifModal.style.display = 'flex';
-  });
 }
 
 const exportBtn = document.getElementById('exportGifBtn');
